[MOV] odoo,auth_oauth: fragment_to_query_string
The aim of this commit is to allow module unrelated to auth_oauth to
import fragment_to_query_string.

Task-id: 6071808
Part-of: odoo/odoo#270834
Related: odoo/enterprise#121019
